LayoutTests/imported/w3c/web-platform-tests/css/css-cascade/idlharness-expected.txt


PASS idl_test setup
PASS idl_test validation
PASS CSSLayerBlockRule interface: existence and properties of interface object
PASS CSSLayerBlockRule interface object length
PASS CSSLayerBlockRule interface object name
PASS CSSLayerBlockRule interface: existence and properties of interface prototype object
PASS CSSLayerBlockRule interface: existence and properties of interface prototype object's "constructor" property
PASS CSSLayerBlockRule interface: existence and properties of interface prototype object's @@unscopables property
PASS CSSLayerBlockRule interface: attribute name
PASS CSSLayerBlockRule must be primary interface of block
PASS Stringification of block
PASS CSSLayerBlockRule interface: block must inherit property "name" with the proper type
PASS CSSLayerStatementRule interface: existence and properties of interface object
PASS CSSLayerStatementRule interface object length
PASS CSSLayerStatementRule interface object name
PASS CSSLayerStatementRule interface: existence and properties of interface prototype object
PASS CSSLayerStatementRule interface: existence and properties of interface prototype object's "constructor" property
PASS CSSLayerStatementRule interface: existence and properties of interface prototype object's @@unscopables property
PASS CSSLayerStatementRule interface: attribute nameList
PASS CSSLayerStatementRule must be primary interface of statement
PASS Stringification of statement
PASS CSSLayerStatementRule interface: statement must inherit property "nameList" with the proper type
FAIL CSSScopeRule interface: existence and properties of interface object assert_own_property: self does not have own property "CSSScopeRule" expected property "CSSScopeRule" missing
FAIL CSSScopeRule interface object length assert_own_property: self does not have own property "CSSScopeRule" expected property "CSSScopeRule" missing
FAIL CSSScopeRule interface object name assert_own_property: self does not have own property "CSSScopeRule" expected property "CSSScopeRule" missing
FAIL CSSScopeRule interface: existence and properties of interface prototype object assert_own_property: self does not have own property "CSSScopeRule" expected property "CSSScopeRule" missing
FAIL CSSScopeRule interface: existence and properties of interface prototype object's "constructor" property assert_own_property: self does not have own property "CSSScopeRule" expected property "CSSScopeRule" missing
FAIL CSSScopeRule interface: existence and properties of interface prototype object's @@unscopables property assert_own_property: self does not have own property "CSSScopeRule" expected property "CSSScopeRule" missing
FAIL CSSScopeRule interface: attribute start assert_own_property: self does not have own property "CSSScopeRule" expected property "CSSScopeRule" missing
FAIL CSSScopeRule interface: attribute end assert_own_property: self does not have own property "CSSScopeRule" expected property "CSSScopeRule" missing
FAIL Stringification of scope assert_class_string: class string of scope expected "[object CSSScopeRule]" but got "[object Null]"
FAIL CSSScopeRule interface: scope must inherit property "start" with the proper type assert_inherits: provided value is not an object
FAIL CSSScopeRule interface: scope must inherit property "end" with the proper type assert_inherits: provided value is not an object
